Rename feature
to `submission-via-ctrl-enter-everywere`
